refactor(home): adding loading and changing styles
