+++ client/src/pages/Home/Home.module.scss
@@ -1,8 +1,7 @@
-.title {
-  margin-top: 8px;
-  margin-bottom: 0;
+.container {
+  padding: 24px;
 }
 
-.subtitle {
-  margin-top: 0;
+.text {
+  margin: 0;
 }
